+++ backend/backend/settings.py
@@ -27,6 +27,9 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 # SECURITY WARNING: keep the secret key used in production secret!
 SECRET_KEY = os.getenv('SECRET_KEY')
 
+# OpenAI API key
+OPENAI_API_KEY = os.getenv('OPENAI_API_KEY')
+
 
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = os.getenv('DEBUG') == 'True'
@@ -44,6 +47,9 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'core',
+    'hobbies',
+    'recommendations',
+    'challenges',
     'rest_framework',
     'rest_framework_simplejwt',
 ]
