+++ backend/core/urls.py
@@ -1,0 +1,11 @@
+from django.urls import path
+from rest_framework_simplejwt.views import TokenRefreshView
+from .views import RegisterView, UserDetailView
+from rest_framework_simplejwt.views import TokenObtainPairView
+
+urlpatterns = [
+    path('register/', RegisterView.as_view(), name='auth_register'),
+    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
+    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
+    path('user/', UserDetailView.as_view(), name='user_detail'),
+]
